merge Feat service

See merge request ind/firmware/uros!2

+++ uros/publisher.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <cstdint>
-
 namespace uros {
 
 template <typename TMsg> struct TopicT;
@@ -27,7 +25,7 @@ template <typename TMsg> struct PublisherT : public PublisherBase {
 
 protected:
   Topic *topic_;
-  int16_t seq_ = 0;
+  int seq_ = 0;
 };
 
 } // namespace uros
